Сделал модуляцию
