+++ js/data.js
@@ -1,0 +1,80 @@
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const DESCRIPTION = [
+  'Название 1',
+  'Название 2',
+  'Название 3',
+  'Название 4',
+  'Название 5',
+  'Название 6',
+  'Название 7',
+  'Название 8',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NUMBER_OBJECT = 25;
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+function createRandomIdFromRangeGenerator (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+const generateId = createRandomIdFromRangeGenerator(1, NUMBER_OBJECT);
+const generatePhotoId = createRandomIdFromRangeGenerator(1, NUMBER_OBJECT);
+const commentsId = createRandomIdFromRangeGenerator(1, NUMBER_OBJECT);
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+const createComentsObject = () => ({
+  id: commentsId(),
+  avatar: `img/avatar-${getRandomInteger(1,6)}.svg`,
+  message: getRandomArrayElement(COMMENTS),
+  name: getRandomArrayElement(NAMES)
+});
+
+const createObject = () => ({
+  id: generateId(),
+  url: `photos/${generatePhotoId()}.jpg`,
+  likes: getRandomInteger(15, 200),
+  description: getRandomArrayElement(DESCRIPTION),
+  comments: Array.from({length: getRandomInteger(2,15)}, createComentsObject),
+});
+
+const createArray = Array.from({length: NUMBER_OBJECT}, createObject);
+
+export {createArray};
